+++ NEWS.md
@@ -1,5 +1,8 @@
 # lgr 0.3.1.9000
 
+* Added AppenderSyslog for logging to syslog via 
+  [rsyslog](https://github.com/atheriel/rsyslog) (thanks to atheriel)
+
 
 # lgr 0.3.1